+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/TruffleContext.java
@@ -272,11 +272,11 @@ public final class TruffleContext implements AutoCloseable {
      * to be paused. The future is completed when all active threads are paused. New threads entered
      * after this point are paused immediately after entering until
      * {@link TruffleContext#resume(Future)} is called.
-     * 
+     *
      * @return a future that can be used to wait for the execution to be paused. Also, the future is
      *         used to resume execution by passing it to the {@link TruffleContext#resume(Future)}
      *         method.
-     * 
+     *
      * @since 21.2
      */
     @TruffleBoundary
@@ -292,13 +292,13 @@ public final class TruffleContext implements AutoCloseable {
      * Resume previously paused execution on all threads for this context. The execution will not
      * resume if {@link TruffleContext#pause()} was called multiple times and for some of the other
      * calls resume was not called yet.
-     * 
+     *
      * @param pauseFuture pause future returned by a previous call to
      *            {@link TruffleContext#pause()}.
-     * 
+     *
      * @throws IllegalArgumentException in case the passed pause future was not obtained by a
      *             previous call to {@link TruffleContext#pause()} on this context.
-     * 
+     *
      * @since 21.2
      */
     @TruffleBoundary