+++ truffle/src/com.oracle.truffle.polyglot/src/com/oracle/truffle/polyglot/PolyglotImpl.java
@@ -487,11 +487,12 @@ public final class PolyglotImpl extends AbstractPolyglotImpl {
 
         PolyglotContextImpl context = languageContext.context;
         PolyglotExceptionImpl exceptionImpl;
-        if (context.closed || context.invalid) {
-            exceptionImpl = new PolyglotExceptionImpl(context.engine, context.cancelling || context.cancelled, e);
+        PolyglotContextImpl.State localContextState = context.state;
+        if (localContextState.isInvalidOrClosed()) {
+            exceptionImpl = new PolyglotExceptionImpl(context.engine, localContextState, e);
         } else {
             try {
-                exceptionImpl = new PolyglotExceptionImpl(languageContext.getImpl(), languageContext.context.engine, context.cancelling || context.cancelled,
+                exceptionImpl = new PolyglotExceptionImpl(languageContext.getImpl(), languageContext.context.engine, localContextState,
                                 languageContext, e, true, entered);
             } catch (Throwable t) {
                 /*
@@ -499,7 +500,7 @@ public final class PolyglotImpl extends AbstractPolyglotImpl {
                  * Report the exception as an internal error.
                  */
                 e.addSuppressed(t);
-                exceptionImpl = new PolyglotExceptionImpl(context.engine, context.cancelling || context.cancelled, e);
+                exceptionImpl = new PolyglotExceptionImpl(context.engine, localContextState, e);
             }
         }
         APIAccess access = getInstance().getAPIAccess();
@@ -511,7 +512,7 @@ public final class PolyglotImpl extends AbstractPolyglotImpl {
         PolyglotEngineException.rethrow(e);
 
         APIAccess access = engine.getAPIAccess();
-        PolyglotExceptionImpl exceptionImpl = new PolyglotExceptionImpl(engine, false, e);
+        PolyglotExceptionImpl exceptionImpl = new PolyglotExceptionImpl(engine, null, e);
         return access.newLanguageException(exceptionImpl.getMessage(), getInstance().exceptionDispatch, exceptionImpl);
     }